refactor: directFundingStatus is a a standalone funciton not affiliated with the Channel model
